Many changes to interface....
Moves the control buttons out of the graphic and into the page. These
can now be customized and fire events as needed. A range slider has been
added to the page, but is not hooked into the viewer code yet. Debugging
mode has been turned on to make it easier to see what is going on in the
viewer.