Fixes z layer mismatch when using range slider
The range slider was not updating the tile graphic correctly. This seems
to be due the update function being called, before the state has been
updated. So, we just pass the layer information directly to the tile
drawing code and update the state independently. This seems to fix the
problem without requiring a silly work around.