Updates location of js files to keep build happy.
The watch process was looking for changes in the javascript js
directory. This triggered a build of the code, which changed the contents
of the javascript directory, which.... This makes sure that the grunt
watch process doesn't keep trying to rebuild the javascript add
infinitum by moving the resulting build files to a different location.

There is also the addition of multiple tile sets to the tile viewer,
which should allow for the viewing of multiple slice planes.