Experimental version that uses grayscale as tile source
The tilesources have been changed to use the grayscale api to get the
tiles. This has the benefit of being able to see the tiles that are on
the XZ or YZ planes, which were not generated by the greytiles data set.
It has the disadvantage that it crushes the dvid server and only the
most zoomed in resolution works. Still, it should help figure out the
coordinate system during development.