Fix to get image centering working on plane change
The image was not centering on the correct coordinates after a plane
change, because the image dimensions had not changed before the
centering method was called. The fix is to move the centering method
into an event handler that is triggered by an event that is fired after
the page (plane) has been updated. This lead to another problem with the
event not firing late enough, so there needed to be two event handlers.
The first event stores the details of the coordinates change and the
second event updates the page centering and rests the variable. I HATE
THIS, but it works. Ideally there should be a single event.