Fixed bug with the navigation window offset.
The fix, as suggest by Bill, was to use the dimensions of the tiled
image, which are larger due to the tile sizes not fitting exactly. This
makes sure that the image in the navigation is scaled from an image that
is the same size at full resolution, so the red viewing window box is in
the correct location.