Removed need for hard coded infotype url.
This url was being used to get the dimensions of the volume, when a tile
source was used that didn't provide the dimensions in the meta data,
such as multiscale2d. However, the multiscale2d does provide a "source"
entry in the meta data, which tells us where to look for the information
pertaining to the original volume from which the tiles were created.
This can then be used to perform a secondary query fro the needed
information with the need to hard code anything. Yay for REST
discovery!!